Fix ResourceWarnings when opening many piped processes
When many subprocesses are spawned, fcntl.fcntl starts to raise an OSError
(permission denied). This is not handled in the
PipedCompressionReader/-Writer __init__ methods, so the OSError is
propagated and then the fallback mechanism finally opens the file using
regular gzip.open. However, the Popen object is never closed correctly if
this happens, leading to the ResourceWarning.

This is likely the cause of https://github.com/marcelm/cutadapt/issues/493