Ignore overzealous mypy error
Literal[rb] Literal[wb] etc is not a subtype of str. As such passing any string into a mode is not accepted.
This is very unhelpful. The point of typing is to preempt TypeErrors. Input sanitization should be done
in the function itself. Since the type system does NOT run at runtime, this kind of typing does not add
anything. It only creates extra busywork.