Open output file for piping in binary mode
This avoids an EncodingWarning that was previously raised when opening the
file in text mode. For Popen, we only need something that has a fileno()
method and encoding or whether the file was opened in text mode at all is
ignored anyway.

Encoding and text mode are taken care of separately when writing into the
process.